Refactor search and profile views
