+++ talentmap_api/user_profile/urls/profile.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from talentmap_api.user_profile import views
+from talentmap_api.user_profile.views import profile as views
 from talentmap_api.common.urls import get_retrieve, patch_update, get_list
 
 urlpatterns = [
